+++ packages/utils/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./activity";
-export * from "./inventory";
+export * from "./clan";
 export * from "./dayjsmhq";
+export * from "./inventory";